Move specialized function `collect_files_to_process` to diff.rs
