+++ src/core/objects/mod.rs
@@ -6,7 +6,6 @@ pub mod traits;
 pub mod tree;
 pub mod worktree;
 
-use std::collections::HashSet;
 use std::fs;
 use std::path::Path;
 
@@ -795,32 +794,6 @@ pub(super) fn get_files(
             .into_iter()
             .collect(),
     })
-}
-
-/// Collects all files that need to be processed, based on user-specified files or default paths.
-///
-/// # Parameters
-/// - `files1`: A slice of `FileSource` representing files from the first tree.
-/// - `files2`: A slice of `FileSource` representing files from the second tree.
-/// - `specified_files`: A slice of `String` with specific files to process, if any.
-///
-/// # Returns
-/// A `Vec<String>` containing paths to all files that need processing.
-pub(super) fn collect_files_to_process(
-    files1: &[FileSource],
-    files2: &[FileSource],
-    specified_files: &[String],
-) -> Vec<String> {
-    let mut all_files = HashSet::new();
-
-    if specified_files.is_empty() {
-        all_files.extend(files1.iter().map(FileSource::path));
-        all_files.extend(files2.iter().map(FileSource::path));
-    } else {
-        all_files.extend(specified_files.iter().cloned());
-    }
-
-    all_files.into_iter().collect()
 }
 
 #[cfg(test)]
